refactor(state-store): split createQueryApi into per-entity sub-factories
EGC-539 audit (Finding 6): createQueryApi in scripts/lib/state-store/
queries.js was a single 742-line function preparing ~30 SQL statements
and exposing ~20 methods across 9 unrelated entities (sessions,
skill_runs, skill_versions, decisions, install_state, governance_events,
instincts, events, lessons, patterns) in one closure.

Split into scripts/lib/state-store/queries/<entity>.js, each owning its
own prepared statements, row mappers, input normalizers, and a
create<Entity>Queries(db) factory. queries.js is now a thin composer:
it builds each sub-API and wires up the two genuinely cross-entity
reads (getSessionDetail, getStatus) by calling into the sub-APIs
instead of touching any prepared statement directly. applyDecaySweep's
db.transaction() usage is unaffected -- lessons.js receives the same
db handle every other module does.

Public surface is unchanged: same method names on the object
createQueryApi(db) returns, same module.exports keys (verified against
the three real external consumers: state-store/index.js, and the two
test files that import REINFORCE_DELTA/DECAY_DELTA_PER_WEEK/
DECAY_GRACE_DAYS/ARCHIVE_THRESHOLD/normalizeLimit directly from
queries.js).

Verified: tests/lib/state-store.test.js 15/15,
tests/lib/state-store-concurrency.test.js 1/1 (2 skipped, pre-existing,
unrelated to this change), tests/lib/lessons-decay.test.js 21/21,
tests/lib/install-state-store-sync.test.js 4/4,
tests/stress/normalize-limit.test.js 4/4. No new test file added --
this is a structural split of already-tested code, not new behavior.
