+++ scripts/lib/state-store/queries/shared.js
@@ -1,0 +1,134 @@
+'use strict';
+
+// Helpers shared across the per-entity query modules in this directory
+// (EGC-539 audit, Finding 6 -- split out of the former 742-line createQueryApi).
+
+const SUCCESS_OUTCOMES = new Set(['success', 'succeeded', 'passed']);
+const FAILURE_OUTCOMES = new Set(['failure', 'failed', 'error']);
+
+function normalizeLimit(value, fallback) {
+  if (value === undefined || value === null) {
+    return fallback;
+  }
+
+  const parsed = Number.parseInt(value, 10);
+  if (!Number.isFinite(parsed) || parsed <= 0) {
+    throw new Error(`Invalid limit: ${value}`);
+  }
+
+  return parsed;
+}
+
+function parseJsonColumn(value, fallback) {
+  if (value === null || value === undefined || value === '') {
+    return fallback;
+  }
+
+  try {
+    return JSON.parse(value);
+  } catch (e) {
+    console.warn(`[StateStore] Failed to parse JSON column: ${e.message}`);
+    return fallback;
+  }
+}
+
+function stringifyJson(value, label) {
+  try {
+    return JSON.stringify(value);
+  } catch (error) {
+    throw new Error(`Failed to serialize ${label}: ${error.message}`, { cause: error });
+  }
+}
+
+function classifyOutcome(outcome) {
+  const normalized = String(outcome || '').toLowerCase();
+  if (SUCCESS_OUTCOMES.has(normalized)) {
+    return 'success';
+  }
+
+  if (FAILURE_OUTCOMES.has(normalized)) {
+    return 'failure';
+  }
+
+  return 'unknown';
+}
+
+function toPercent(numerator, denominator) {
+  if (denominator === 0) {
+    return null;
+  }
+
+  return Number(((numerator / denominator) * 100).toFixed(1));
+}
+
+function summarizeSkillRuns(skillRuns) {
+  const summary = {
+    totalCount: skillRuns.length,
+    knownCount: 0,
+    successCount: 0,
+    failureCount: 0,
+    unknownCount: 0,
+    successRate: null,
+    failureRate: null,
+  };
+
+  for (const skillRun of skillRuns) {
+    const classification = classifyOutcome(skillRun.outcome);
+    if (classification === 'success') {
+      summary.successCount += 1;
+      summary.knownCount += 1;
+    } else if (classification === 'failure') {
+      summary.failureCount += 1;
+      summary.knownCount += 1;
+    } else {
+      summary.unknownCount += 1;
+    }
+  }
+
+  summary.successRate = toPercent(summary.successCount, summary.knownCount);
+  summary.failureRate = toPercent(summary.failureCount, summary.knownCount);
+  return summary;
+}
+
+function summarizeInstallHealth(installations) {
+  if (installations.length === 0) {
+    return {
+      status: 'missing',
+      totalCount: 0,
+      healthyCount: 0,
+      warningCount: 0,
+      installations: [],
+    };
+  }
+
+  const summary = installations.reduce((result, installation) => {
+    if (installation.status === 'healthy') {
+      result.healthyCount += 1;
+    } else {
+      result.warningCount += 1;
+    }
+    return result;
+  }, {
+    totalCount: installations.length,
+    healthyCount: 0,
+    warningCount: 0,
+  });
+
+  return {
+    status: summary.warningCount > 0 ? 'warning' : 'healthy',
+    ...summary,
+    installations,
+  };
+}
+
+module.exports = {
+  SUCCESS_OUTCOMES,
+  FAILURE_OUTCOMES,
+  normalizeLimit,
+  parseJsonColumn,
+  stringifyJson,
+  classifyOutcome,
+  toPercent,
+  summarizeSkillRuns,
+  summarizeInstallHealth,
+};
